refs: alternate reftable ref backend implementation
This introduces the reftable backend as an alternative to the packed
backend. This simplifies the code, because we now longer have to worry
about:

- pseudorefs and the HEAD ref
- worktrees
- commands that blur together files and references (cherry-pick, rebase)

This deviates from the spec that in
Documentation/technical/reftable.txt. It might be possible to update
the code such that all writes by default go to reftable directly. Then
the result would be compatible with an implementation that writes only
reftable (the reftable lock would still prevent races) relative to an
implementation that disregards loose refs. Or,  JGit could be adapted
to follow this implementation.

For this incremental path, the reftable format is arguably more
complex than necessary, as

- packed-refs doesn't support symrefs
- reflogs aren't moved into reftable/

on the other hand, the code is already there, and it's well-structured
and well-tested.

This implementation is a prototype. To test, you need to do `export
GIT_TEST_REFTABLE=1`. Doing so creates a handful of errors in the
test-suite, most seemingly related to the new behavior of pack-refs
(which creates a reftable/ dir and not a packed-refs file.), but it
seems overseeable.

Signed-off-by: Han-Wen Nienhuys <hanwen@google.com>